+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased (1.2.7) (XXXX-XX-XX)
 
 - Make backend domain constant in lizard-nxt module.
 
+- Change CNAME for gh-pages.
+
 - Update installation documentation.